test: [#414] download torrent in private mode
The torrent should include the tracker keys.

+++ cypress/e2e/contexts/torrent/specs/public_download.cy.ts
@@ -52,34 +52,36 @@ describe("A guest user", () => {
     cy.delete_user_from_database(uploader_registration_form.username);
   });
 
-  it("should be able to download a preexisting torrent", () => {
-    const torrent_info = generateRandomTestTorrentInfo();
+  if (Cypress.env("TRACKER_MODE") === "public") {
+    it("should be able to download a preexisting torrent", () => {
+      const torrent_info = generateRandomTestTorrentInfo();
 
-    cy.login(uploader_registration_form.username, uploader_registration_form.password);
+      cy.login(uploader_registration_form.username, uploader_registration_form.password);
 
-    cy.upload_torrent(torrent_info);
+      cy.upload_torrent(torrent_info);
 
-    cy.get("div[data-cy=\"user-menu\"]").click();
+      cy.get("div[data-cy=\"user-menu\"]").click();
 
-    cy.logout();
+      cy.logout();
 
-    // Ensure we are still on the torrent details page
-    cy.url().should("include", "/torrent/");
+      // Ensure we are still on the torrent details page
+      cy.url().should("include", "/torrent/");
 
-    cy.intercept({
-      method: "GET",
-      url: "/*/torrent/download/*"
-    }).as("download");
+      cy.intercept({
+        method: "GET",
+        url: "/*/torrent/download/*"
+      }).as("download");
 
-    cy.get("button[data-cy=\"torrent-action-download\"]").click();
+      cy.get("button[data-cy=\"torrent-action-download\"]").click();
 
-    cy.wait("@download").then((interception) => {
-      // Ensure the filename is correct
-      expect(interception.response.headers["content-disposition"]).to.include(torrent_info.filename);
+      cy.wait("@download").then((interception) => {
+        // Ensure the filename is correct
+        expect(interception.response.headers["content-disposition"]).to.include(torrent_info.filename);
 
-      // Delete the test torrent generated for this test
-      const torrentInfoHash = parseInfoHash(interception.response.headers["x-torrust-torrent-infohash"]);
-      cy.delete_torrent_from_database_and_fixture(torrent_info, torrentInfoHash);
+        // Delete the test torrent generated for this test
+        const torrentInfoHash = parseInfoHash(interception.response.headers["x-torrust-torrent-infohash"]);
+        cy.delete_torrent_from_database_and_fixture(torrent_info, torrentInfoHash);
+      });
     });
-  });
+  }
 });
